Shift Frame() out of accounting into new ethernet module
This allows it to be shared between the ceilometer client and the
web aggregator if that should ever eventuate.
